requirements.txt
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
flake8==7.0.0
mccabe==0.7.0
pillow==10.2.0
pycodestyle==2.11.1
pyflakes==3.2.0
sqlparse==0.4.4
typing_extensions==4.10.0
mysqlclient==2.2.4
django-redis==5.4.0
redis==5.0.4
requests
django-extensions
django-select2
django-import-export
celery
django-celery-results
django-celery-beat
django-prometheus
python-logging-loki
python-dotenv
uWSGI
djangorestframework-simplejwt
django-cors-headers
django-allauth
dj-rest-auth
cryptography
django-ratelimit
